+++ src/components/heading/__spec__.js
@@ -102,4 +102,14 @@ describe('Heading', () => {
       expect(TestUtils.scryRenderedComponentsWithType(instance, Link).length).toEqual(0);
     });
   });
+
+  describe('with separator', () => {
+    it('renders a separator after the title', () => {
+      instance = TestUtils.renderIntoDocument(
+        <Heading title="foo" separator={ true }/>
+      );
+      let separator = TestUtils.findRenderedDOMComponentWithTag(instance, 'hr');
+      expect(separator.props.className).toEqual('carbon-heading__separator');
+    });
+  });
 });