+++ CHANGELOG.md
@@ -16,6 +16,10 @@ We have added a settings row component for settings pages. It employs the curren
 </SettingsRow>
 ```
 
+## Component Enhancements
+
+* `Heading`: One new prop has been added, `separator`, to show a 2x50px separator between title and subheader.
+
 # 0.29.1
 
 ## CSS Update